+++ src/domain/entities/LsifDataEntities.ts
