+++ src/domain/errors/AdditionalErrorInfo.ts
