+++ src/domain/errors/ErrorStatus.ts
