+++ src/domain/logs/CreateLoggerFactory.ts
@@ -1,6 +1,6 @@
-import createServerLogger from 'domain/logs/serverLogger';
-import createBrowserLogger from 'domain/logs/browserLogger';
-import { BaseLogger } from 'domain/logs/loggerInterface';
+import createServerLogger from 'domain/logs/ServerLogger';
+import createBrowserLogger from 'domain/logs/BrowserLogger';
+import { BaseLogger } from 'domain/logs/LoggerInterface';
 
 const createLogger = (): BaseLogger => {
   if (typeof window === 'undefined') {
