+++ src/domain/logs/LoggerInterface.ts
